% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{abundance.matrix}
\alias{abundance.matrix}
\title{abundance matrix}
\usage{
abundance.matrix(data, times = seq(0, data$maxtime, length.out = 50),
  by.age = F, ...)
}
\arguments{
\item{data}{result of a simulation, created by \code{\link{community}}}

\item{times}{array of times at which the abundances will be calculated}

\item{by.age}{T/F. Use this option to get the number of individuals to reach each age, instead of
abundances for each time.}

\item{\dots}{additional parameters to be passed to the \code{\link{age.data}} function}
}
\description{
creates a matrix with abundances of each life stage/species over time
}
\examples{
malth <- community(2,3,param,dispersal=2,init=c(100,0,0,0))
times <- seq(0,2,by=0.1)
ab <- abundance.matrix(malth,times)
stackplot(ab[,1:3])
}
