% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{abundance_matrix}
\alias{abundance_matrix}
\title{creates a matrix with abundances of each life stage/species over time}
\usage{
abundance_matrix(data, times = seq(0, data$maxtime, length.ou = 20))
}
\arguments{
\item{data}{result of a simulation, created by \code{\link{facilitation}}}

\item{times}{array of times at which the abundances will be calculated}
}
\description{
creates a matrix with abundances of each life stage/species over time
}
\examples{
malth <- facilitation(2,3,Ds=c(5,1.2,0.1),Gs=c(1,.5),R=10,dispersal=2,init=c(100,0,0,0))
times <- seq(0,2,by=0.1)
ab <- abundance_matrix(malth,times)
stackplot(ab[,1:3])
}
