% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{age.data}
\alias{age.data}
\title{lifehistory relative to birthdate of each individual (used by abundance.matrix with by.age)}
\usage{
age.data(data, cap.living = F)
}
\arguments{
\item{data}{result of a simulation, created by \code{\link{community}}}

\item{cap.living}{T/F. Use this option to set the time of death of living individuals to max
simulation time. Otherwise, living individuals are excluded from the data. Either way, this data
will be more representative if only a small fraction of total individuals is living at the end of
simulation.}
}
\description{
lifehistory relative to birthdate of each individual (used by abundance.matrix with by.age)
}
