% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{community}
\alias{community}
\title{community}
\usage{
community(maxtime, numstages, parameters, init, interactionsD, interactionsG,
  interactionsR, height = 100, width = 100, boundary = c("reflexive",
  "absortive", "periodic"), dispKernel = c("exponential", "random"),
  maxpop = 30000)
}
\arguments{
\item{maxtime}{How long the simulation must run}

\item{numstages}{Array of number of stages for each population}

\item{parameters}{Data.frame or matrix with one row for each stage. Columns:
D,G,R,dispersal distance,radius(optional),maxstressefect (optional)}

\item{init}{Either an array of initial numbers for each stage of each population, or a
data.frame with the history of a simulation}

\item{interactionsD}{Optional. A square matrix of effects of life stages over each other, where element
[i,j] is the effect of stage i over stage j. Positive values equal facilitation, negative
ones, competition. The interactions occur only if the affected individual is within the affecting
individual's radius, and are additive. Affects death rates (is subtracted from D).}

\item{interactionsG}{Same as above, but affecting growth rates (is added to G).}

\item{interactionsR}{Same as above, but affecting reproduction rates (is added to R) .}

\item{height}{Arena height}

\item{width}{Arena width}

\item{boundary}{Type of boundary condition. Options are "reflexive", "absortive" and
"periodic". Default is reflexive.}

\item{dispKernel}{Type of dispersion kernel. Options are "exponential" and "random", in which
seeds are dispersed randomly regardless of parent position (note: "random" option ignores
dispersal parameter)}

\item{maxpop}{If the simulation reaches this many individuals total, it will stop. Default
is 30000.}
}
\description{
Runs a simulation with any number of structured populations, for a limited time.
}
\examples{
init <- list(c(100,0,0),c(100,0))
###               D G R d  D G R d ...
param <- matrix(c(2,1,0,0, 1,1,0,0, .5,0,6,2, 1,1,0,0, .5,0,2,2), byrow=TRUE, nrow=5) 
malth <- community(3,c(3,2),param,init=init)
times <- seq(0,3,by=0.1)
ab <- abundance.matrix(malth,times)
stackplot(ab[,1:3]) # species 1
stackplot(ab[,4:5]) # species 2
}
