% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{longevity}
\alias{longevity}
\title{longevity}
\usage{
longevity(data)
}
\arguments{
\item{data}{result of a simulation, created by \code{\link{community}}}
}
\description{
calculates the lifespan of each individual
}
\examples{
malth <- community(2,3,Ds=c(5,1.2,0.1),Gs=c(1,.5),R=10,dispersal=2,init=c(100,0,0,0))
l <- longevity(malth)
hist(l$longevity)
}
