% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mat_model.R
\name{mat_model}
\alias{mat_model}
\title{mat_model}
\usage{
mat_model(n = 3, Ds = rexp(n, 1), Gs = rexp(n - 1, 1), R = rexp(1, 1))
}
\arguments{
\item{n}{The number of life stages. Default is 3.}

\item{Ds}{An n-array with death rates for each stage. Default is rexp(n,1).}

\item{Gs}{An (n-1)-array with growth rates for each stage but the last. Default is rexp(n-1,1)}

\item{R}{Reproduction rate. Default is rexp(1,1).}
}
\description{
The \code{mat_model} function produces the interaction matrix for a structured
population model where only the last life stage reproduces, to be applyed in a linear ODE
}
\examples{
mat <- mat_model(5)
mat2 <- mat_model(3,c(1,2,3),c(10,10),100)
}

