% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{restart}
\alias{restart}
\title{restart}
\usage{
restart(data, time, start = 0)
}
\arguments{
\item{data}{result of a simulation, created by \code{\link{community}}}

\item{time}{a number: for how long to extend the simulation}

\item{start}{a number: an instant in time to begin from}
}
\description{
Turn back time and restart a simulation from time t
}
