% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_model.R
\name{simulation.to.mat.models}
\alias{simulation.to.mat.models}
\title{Generate a matrix population model from the simulation parameters}
\usage{
simulation.to.mat.models(data)
}
\arguments{
\item{data}{the result of a simulation}
}
\description{
Generate a matrix population model from the simulation parameters
}
