% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_model.R
\name{solution.matrix}
\alias{solution.matrix}
\title{solution.matrix}
\usage{
solution.matrix(p0, M, times = c(1:10))
}
\arguments{
\item{p0}{initial condition, as an array}

\item{M}{a square matrix with as many rows as P0}

\item{times}{an array containing the times in which to calculate the solution}
}
\description{
The \code{solution.matrix} function returns the solution to a linear ODE of the form P' = MP,
which is merely P(t) = exp(Mt)p0 where p0 is the initial condition
}
\examples{
mat <- mat.model.base(5)
solution.matrix(c(1,0,0,0,0),mat)
}
