% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{spatialanimation}
\alias{spatialanimation}
\alias{plotsnapshot}
\title{function for ploting simulation frames}
\usage{
spatialanimation(data, times = seq(0, data$maxtime, length.out = 50),
  interval = 0.1, movie.name = "facilitationmovie.gif", xlim = c(0,
  data$w), ylim = c(0, data$h), color = c(colorRampPalette(c("darkred",
  "pink"))(data$num.total - 1), "lightgreen"), ...)

plotsnapshot(data, t, ...)
}
\arguments{
\item{data}{result of a simulation, created by \code{\link{facilitation}}}

\item{times}{array of times at which to plot}

\item{xlim}{Optional. Limits to the x-axis}

\item{ylim}{Optional. Limits to the y-axis}

\item{color}{Optional. A color vector}

\item{t}{a single time at which to plot}

\item{tframe}{a time length to wait between frames. Do not use if using this with
\code{animation}}
}
\description{
function for ploting simulation frames

make a gif with a single frame
}
\examples{
malth <- facilitation(2,3,c(5,1,.1),c(1,.5),10,dispersal=2,init=c(100,0,0,0),radius=c(0,1,2,0))
times <- seq(0,2,by=0.1)
# plot
spatialanimation(malth,times,interval=.1,movie.name="malthusian.gif")
}
\author{
Alexandre Adalardo de Oliveira - 16/03/2016

M. Salles
}
