% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{stackplot}
\alias{stackplot}
\title{Plots}
\usage{
stackplot(mat, col, ...)
}
\arguments{
\item{mat}{A population matrix, as produced by \code{\link{abundance_matrix}}}

\item{col}{Optional. A color vector}

\item{\dots}{Further parameters to be passed to the lower leve plot}
}
\description{
The \code{stackplot} function produces a stacked plot of the population over time.
}
\examples{
obj <- test_standard(10, 10)
mat <- abundance_matrix(obj)
stackplot(mat)
}

