% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_model.R
\name{limiting.rate}
\alias{limiting.rate}
\title{a function to calculate the dominant eigenvalue of a matrix}
\usage{
limiting.rate(mat)
}
\arguments{
\item{mat}{a matrix}
}
\description{
a function to calculate the dominant eigenvalue of a matrix
}
