% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{longevity}
\alias{longevity}
\title{longevity}
\usage{
longevity(data)
}
\arguments{
\item{data}{result of a simulation, created by \code{\link{community}}}
}
\description{
Calculates the lifespan of each individual. Returns a data.frame with the individual's id,
the last stage reached by that individual, the time of birth, time of death (if dead), and
longevity (if dead).
}
\examples{
data(malthusian)
longevity(malthusian)
}
