% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{longevity}
\alias{longevity}
\title{calculates the lifespan of each individual}
\usage{
longevity(data)
}
\arguments{
\item{data}{result of a simulation, created by \code{\link{community}}}
}
\description{
calculates the lifespan of each individual
}
