% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{malthusian}
\alias{malthusian}
\title{A single species, 3 stages, no interactions simulation result}
\format{A list of data and parameters, generated by \link{community}}
\usage{
malthusian
}
\description{
A single species, 3 stages, no interactions simulation result
}
\examples{
## Simple mathusian one species
init <- c(0,0,100)
################# D G R disp
param <- matrix(c(5,1,0,5, 1,1,0,5, .5,0,10,5),nrow=3,byrow=TRUE)
malthusian <- community(10,3,param,init)
}
\keyword{datasets}
