% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_model.R
\name{mat.model}
\alias{mat.model}
\title{matrix population model}
\usage{
mat.model(data, ns, combine.matrices = F)
}
\arguments{
\item{data}{Either the result of a simulation, to extract the parameters from (obs: may return a list of
matrices), or a data.frame containing the parameters}

\item{ns}{an array of numbers of stages. Use for more than one population.}

\item{combine.matrices}{Logical. Combine the matrices into a single, multi-population matrix?}
}
\description{
Produces the mpm matrix for a continuous time structured
population model, to be applied in a linear ODE
}
\examples{
mat <- mat.model(malth)
}
