% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_model.R
\name{mat.model}
\alias{mat.model}
\title{mat.model}
\usage{
mat.model(n = 3, Ds = runif(n, rep(1e-05, n), c(rep(2, n - 1), 0.01)),
  Gs = runif(n - 1, 1e-05, 2), R = runif(1, 1e-04, 10))
}
\arguments{
\item{n}{The number of life stages. Default is 3.}

\item{Ds}{An n-array with death rates for each stage.}

\item{Gs}{An (n-1)-array with growth rates for each stage but the last.}

\item{R}{Reproduction rate. Default is rexp(1,1).}
}
\description{
The \code{mat.model} function produces the interaction matrix for a structured
population model where only the last life stage reproduces, to be applyed in a linear ODE
}
\examples{
mat <- mat.model(5)
mat2 <- mat.model(3,c(1,2,3),c(10,10),100)
}
