% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{proceed}
\alias{proceed}
\title{proceed}
\usage{
proceed(data, time)
}
\arguments{
\item{data}{result of a simulation, created by \code{\link{community}}}

\item{time}{a number: for how long to extend the simulation}
}
\description{
Proceed with a stopped simulation.
}
