% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{proceed}
\alias{proceed}
\title{proceed with a stopped simulation}
\usage{
proceed(data, time)
}
\arguments{
\item{data}{result of a simulation, created by \code{\link{facilitation}}}

\item{time}{a number: for how long to extend the simulation}
}
\description{
proceed with a stopped simulation
}
