% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{spatialplot}
\alias{spatialplot}
\title{function for ploting simulation frames}
\usage{
spatialplot(dtlist, times = 1:length(data), xlim, ylim, sp,
  col = colorRampPalette(c("darkred", "pink"))(length(sp)), radius)
}
\arguments{
\item{dtlist}{A list of data.frames, one for each time in times, containing columns sp,x,y}

\item{times}{Array of times, corresponding to the data.frames}

\item{xlim}{Limits to the x-axis}

\item{ylim}{Limits to the y-axis}

\item{sp}{Array of species/stages id, in order of plotting bottom to top}

\item{col}{A color array, one for each species/stage}

\item{radius}{Array or radiuses, one for each species/stage}
}
\description{
function for ploting simulation frames
}
\author{
Alexandre Adalardo de Oliveira - 16/03/2016

M. Salles
}
