% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{stackplot}
\alias{stackplot}
\title{Plots}
\usage{
stackplot(mat, col, legend, log.y = FALSE, ...)
}
\arguments{
\item{mat}{A population matrix, as produced by \code{\link{abundance_matrix}}}

\item{col}{Optional. A color vector}

\item{legend}{Optional. An array of names}

\item{log.y}{Logical. Should the y-axis be plotted in a logarithmic scale?}

\item{\dots}{Further parameters to be passed to the lower level plot function}
}
\description{
Plotting functions.
}
\details{
The \code{stackplot} function produces a stacked plot of the population over time.
Notice that the population should have at least two stages for this function to work.
}
\examples{
obj <- facilitation(maxtime=2,n=3,Ds=c(5,1.2,0.1),Gs=c(1,.5),R=10,dispersal=2,init=c(100,0,0,0))
times <- seq(0,2,by=0.1)
ab <- abundance_matrix(obj,times)
stackplot(ab[,1:3])
}
